// apps/react-frontend/src/api/client/models/sync-preview-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Eazy Recycling Backend Service
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ExternalProductToDeleteResponse } from './external-product-to-delete-response';
// May contain unused imports in some cases
// @ts-ignore
import type { MaterialToSyncResponse } from './material-to-sync-response';
// May contain unused imports in some cases
// @ts-ignore
import type { SyncPreviewSummary } from './sync-preview-summary';

/**
 * 
 * @export
 * @interface SyncPreviewResponse
 */
export interface SyncPreviewResponse {
    /**
     * 
     * @type {Array<MaterialToSyncResponse>}
     * @memberof SyncPreviewResponse
     */
    'toCreate': Array<MaterialToSyncResponse>;
    /**
     * 
     * @type {Array<MaterialToSyncResponse>}
     * @memberof SyncPreviewResponse
     */
    'toUpdate': Array<MaterialToSyncResponse>;
    /**
     * 
     * @type {Array<ExternalProductToDeleteResponse>}
     * @memberof SyncPreviewResponse
     */
    'toDelete': Array<ExternalProductToDeleteResponse>;
    /**
     * 
     * @type {Array<MaterialToSyncResponse>}
     * @memberof SyncPreviewResponse
     */
    'unchanged': Array<MaterialToSyncResponse>;
    /**
     * 
     * @type {SyncPreviewSummary}
     * @memberof SyncPreviewResponse
     */
    'summary': SyncPreviewSummary;
}